requirements.txt
# MCP31PrinterBOT - プリンタ受付サーバー用依存関係

# 画像処理
pillow>=11.0.0

# プリンター制御 (ESC-POS)
python-escpos>=3.0

# 管理Web API
flask>=3.0.0
